Updated navbar implementation
- combined top and bottom navbar
- removed comments
- made it rendered in app.js
