+++ src/Containers/Pages/MainRecipeModal/MainRecipeModal.module.scss
@@ -1,8 +1,8 @@
 .fullPage {
   text-align: center;
   position: absolute;
-  height: 100%;
-  bottom: 0;
+  height: 75%;
+  bottom: 10%;
 }
 
 .recipeName {
