+++ src/Containers/Pages/ProfilePage/ProfilePage.module.scss
@@ -13,7 +13,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  padding-top: 42px;
+  padding-top: 40%;
   padding-bottom: 14px;
   width: 100%;
   border-bottom: 0.3px solid $preferenceSettingColor;
