+++ src/component/navbars/Navbar.module.scss
@@ -1,10 +1,9 @@
 .bottomNavContainer {
     position: absolute;
-    bottom: 5%;
+    bottom: 0%;
     display: flex;
     flex-direction: row;
     width: 100%;
-    // overflow: hidden;
     border-top: 0.3px solid #DB5A42;
     text-align: center;
   
@@ -19,11 +18,11 @@
 }
 
 .topNavContainer {
-    //position: absolute;
-    top: 5%;
+    position: absolute;
+    top: 0%;
     display: flex;
     flex-direction: row;
-    //width: 90%;
+    width: 100%;
     overflow: hidden;
     border-bottom: 0.3px solid #DB5A42;
     text-align: center;
